requirements.txt
django==4.2.6
django-cors-headers==4.3.0
djangorestframework==3.14.0